refactor osc
