+++ lib/avatars_page.dart
@@ -1,6 +1,5 @@
 import 'dart:async';
 import 'dart:convert';
-import 'dart:io';
 import 'package:collection/collection.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
@@ -17,6 +16,7 @@ import 'package:vrc_avatar_manager/order_dialog.dart';
 import 'package:vrc_avatar_manager/performance_selector.dart';
 import 'package:vrc_avatar_manager/prefs.dart';
 import 'package:vrc_avatar_manager/setting_dialog.dart';
+import 'package:vrc_avatar_manager/vrcosc.dart';
 import 'package:vrc_avatar_manager/wrap_with_height.dart';
 import 'package:vrc_avatar_manager/sort_by.dart';
 import 'package:vrc_avatar_manager/tag_button.dart';
@@ -25,7 +25,6 @@ import 'package:vrc_avatar_manager/tag_edit_dialog.dart';
 import 'package:vrc_avatar_manager/vrc_api.dart';
 import 'package:vrc_avatar_manager/vrc_icons.dart';
 import 'package:vrchat_dart/vrchat_dart.dart';
-import 'package:udp/udp.dart';
 
 class AvatarsPage extends StatefulWidget {
   const AvatarsPage({super.key, required this.accountId});
@@ -393,40 +392,8 @@ class _AvatarsPageState extends State<AvatarsPage> {
   }
 
   Future<void> _doChangeAvatarOSC(String id) async {
-    const String vrchatOscAddress = "127.0.0.1";
-    const int vrchatOscPort = 9000;
-    const String oscPath = "/avatar/change";
-
-    // OSCメッセージのフォーマット（"/avatar/change" + id のOSCパケット）
-    List<int> messageBytes = _buildOscMessage(oscPath, id);
-
-    // UDPソケットを作成して送信
-    UDP sender = await UDP.bind(Endpoint.any());
-    await sender.send(
-        Uint8List.fromList(messageBytes),
-        Endpoint.unicast(InternetAddress(vrchatOscAddress),
-            port: Port(vrchatOscPort)));
-
-    print("Sent OSC message to VRChat: $oscPath $id");
-
-    sender.close();
-  }
-
-  // OSCメッセージのフォーマットを作成する関数
-  List<int> _buildOscMessage(String address, String argument) {
-    List<int> addressBytes = _oscStringToBytes(address);
-    List<int> typeTagBytes = _oscStringToBytes(",s"); // OSCの型タグ（文字列: "s"）
-    List<int> argumentBytes = _oscStringToBytes(argument);
-    return [...addressBytes, ...typeTagBytes, ...argumentBytes];
-  }
-
-// OSCの文字列をバイト配列に変換する関数（Null文字で終端し、4バイト区切り）
-  List<int> _oscStringToBytes(String value) {
-    List<int> bytes = value.codeUnits + [0]; // 文字列にNull終端を追加
-    while (bytes.length % 4 != 0) {
-      bytes.add(0); // 4バイト境界に揃える
-    }
-    return bytes;
+    await VRCOSC().sendAvatar(id);
+    _showInfo("Avatar changed");
   }
 
   void _toggleTagAvatar(String id) async {
